chem_mcp_toolkit/pyproject.toml
[project]
name = "chem-mcp-toolkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "litellm>=1.67.4",
    "mcp[cli]>=1.6.0",
    "molbloom>=2.3.4",
    "pandas>=2.2.3",
    "pubchempy>=1.0.4",
    "rdchiral>=1.1.0",
    "rdkit>=2024.9.6",
    "requests>=2.32.3",
    "rxn4chemistry>=1.14.0",
    "selfies>=2.2.0",
    "tavily-python>=0.7.0",
    "transformers>=4.51.3",
]
